fix: files

+++ vps/scripts/restore_backups.sh
@@ -1,34 +1,39 @@
 #!/usr/bin/bash
 BUCKET_NAME="stuff"
 PLAUSIBLE_DB_CONTAINER=plausible-plausible_db-1
+# this check is to see if we are only restoring files and not dbs
+ONLY_RESTORE_FILES=false
+if [[ "$1" == "true" ]]; then
+    ONLY_RESTORE_FILES=true
+fi
 
 # Find the most recent backup
 MOST_RECENT_BACKUP=$(rclone lsf "r2:$BUCKET_NAME/backups" --format "p" | sort | tail -n 1)
 BACKUP_FILES=$(rclone lsf "r2:$BUCKET_NAME/backups/$MOST_RECENT_BACKUP" --format "p")
 
-SQL_ARCHIVES=$(echo $BACKUP_FILES | tr ' ' '\n' | grep "\.sql\.zst$")
-for SQL_ARCHIVE in $SQL_ARCHIVES; do
-    IS_PLAUSIBLE=false
-    if [[ $SQL_ARCHIVE == "plausible*" ]]; then
-        IS_PLAUSIBLE=true
-    fi
-    echo "Restoring database from $SQL_ARCHIVE"
-    # Extract SQL archive
-    rclone copyto "r2:$BUCKET_NAME/backups/$MOST_RECENT_BACKUP/$SQL_ARCHIVE" /tmp/$SQL_ARCHIVE
+if [ ! $ONLY_RESTORE_FILES ]; then
+    SQL_ARCHIVES=$(echo $BACKUP_FILES | tr ' ' '\n' | grep "\.sql\.zst$")
+    for SQL_ARCHIVE in $SQL_ARCHIVES; do
+        echo "Restoring database from $SQL_ARCHIVE"
+        # Extract SQL archive
+        rclone copyto "r2:$BUCKET_NAME/backups/$MOST_RECENT_BACKUP/$SQL_ARCHIVE" /tmp/$SQL_ARCHIVE
 
-    echo "Extracting SQL archive"
-    SQL_FILENAME="/tmp/${SQL_ARCHIVE%.zst}"
-    zstd -fd /tmp/$SQL_ARCHIVE -o $SQL_FILENAME
-    rm -f /tmp/$SQL_ARCHIVE
+        echo "Extracting SQL archive"
+        SQL_FILENAME="/tmp/${SQL_ARCHIVE%.zst}"
+        zstd -fd /tmp/$SQL_ARCHIVE -o $SQL_FILENAME
+        rm -f /tmp/$SQL_ARCHIVE
 
-    echo "Restoring database from $SQL_FILENAME"
-    # Restore the full database
-    if [ "$IS_PLAUSIBLE" = true ]; then
-        sudo docker exec -i $PLAUSIBLE_DB_CONTAINER psql -U postgres -d postgres < $SQL_FILENAME
-    fi
-    sudo docker exec -i db psql -U postgres -d postgres < $SQL_FILENAME
-    rm -f $SQL_FILENAME
-done
+        echo "Restoring database from $SQL_FILENAME"
+        # Restore the full database
+        if [[ $SQL_ARCHIVE == "plausible*" ]]; then
+            sudo docker exec -i $PLAUSIBLE_DB_CONTAINER psql -U postgres -d postgres < $SQL_FILENAME
+            rm -f $SQL_FILENAME
+            continue
+        fi
+        sudo docker exec -i db psql -U postgres -d postgres < $SQL_FILENAME
+        rm -f $SQL_FILENAME
+    done
+fi
 
 VAULTWARDEN_ARCHIVE=$(echo $BACKUP_FILES | tr ' ' '\n' | grep "^vaultwarden-data")
 if [ -n "$VAULTWARDEN_ARCHIVE" ]; then
